+++ apps/src/componentLibrary/button/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.0]()
+
+* added background fill color to secondary black and white buttons
+* added ```gray``` secondary button
+* deprecated ```purple``` secondary button
+
 ## [0.3.0](https://github.com/code-dot-org/code-dot-org/pull/58469)
 
 * added support of all `aria-` attributes by `Button` components